pyproject.toml
[project]
name = "helix-mind"
version = "0.1.0"
description = "Helix-Mind: Memory Microservice for Helix Ecosystem"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "duckdb>=1.0",
    "pyarrow>=18.0",
    "sentence-transformers>=3.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "structlog>=24.0",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "gitpython>=3.1",
    "python-frontmatter>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-benchmark>=4.0",
    "ruff>=0.11",
    "mypy>=1.0",
]

[project.scripts]
mind = "mind.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
